+++ demo/Pods/Target Support Files/Pods-SuperAwesome DemoTests/Pods-SuperAwesome DemoTests-resources.sh
@@ -63,10 +63,8 @@ if [[ "$CONFIGURATION" == "Debug" ]]; then
   install_resource "../../SuperAwesome/Resources/displayRichMedia.html"
   install_resource "../../SuperAwesome/Resources/displayTag.html"
   install_resource "../../SuperAwesome/Resources/mark.png"
-  install_resource "../../SuperAwesome/Resources/sa_padlock.png"
-  install_resource "../../SuperAwesome/Resources/SAPadlockView.xib"
-  install_resource "../../SuperAwesome/Resources/start_notselected.png"
-  install_resource "../../SuperAwesome/Resources/start_selected.png"
+  install_resource "../../SuperAwesome/Resources/watermark_49x25.png"
+  install_resource "../../SuperAwesome/Resources/watermark_67x25.png"
 fi
 if [[ "$CONFIGURATION" == "Release" ]]; then
   install_resource "../../SuperAwesome/Resources/close.png"
@@ -74,10 +72,8 @@ if [[ "$CONFIGURATION" == "Release" ]]; then
   install_resource "../../SuperAwesome/Resources/displayRichMedia.html"
   install_resource "../../SuperAwesome/Resources/displayTag.html"
   install_resource "../../SuperAwesome/Resources/mark.png"
-  install_resource "../../SuperAwesome/Resources/sa_padlock.png"
-  install_resource "../../SuperAwesome/Resources/SAPadlockView.xib"
-  install_resource "../../SuperAwesome/Resources/start_notselected.png"
-  install_resource "../../SuperAwesome/Resources/start_selected.png"
+  install_resource "../../SuperAwesome/Resources/watermark_49x25.png"
+  install_resource "../../SuperAwesome/Resources/watermark_67x25.png"
 fi
 
 mkdir -p "${CONFIGURATION_BUILD_DIR}/${UNLOCALIZED_RESOURCES_FOLDER_PATH}"